Found some errors in the initial testing of the previous commit.
After update of a product while it was filtered, caused a sql
error.
The combination attribute generation had the potential to
create a combination that could not be purchased (items were not
sorted)
A SQL statement was incorrectly pieced together incorrectly using
an Execute statement instead of a bindVars statement.
typo/oversight while converting.