Update save button in products_with_attributes_stock.php
Haven't tested; however, appears that the form was not fully supported from top to bottom after my modification.  The above change should restore the functionality of the save button at the bottom of the form, but overall need to look at the effect as well... Currently the $ReturnedProductID does not appear to be set even when the product is filtered.  As for modifications from the main admin page, it may be that a save button needs to be applied to each product entry in order to make the change(s) there when all product is listed so that changes to individual product are captured or some variation to allow all changes on the page to be saved.  More to follow.