Modified the presentation of customid on admin/orders
One change to the presentation is that previously any
single customid assigned to an attribute of a mixed
attribute would be displayed.  It is now set to display
a customid only for a matched SBA variant.
If this is not as expected, then probably need to provide
an option to pick which way to work this.