Modified includes/modules/YOUR_TEMPLATE/attributes.php
Adjusted the display of multiple-attributes so that each
attribute option value appears one time in each option name
instead of the quantity of values in one option name times the
quantity of valus assigned to each of the other option names.
In otherwords, duplication of optino values has been removed;
however so has the ability to display the quantity avail at
the moment.  Ie... Functional without Dynamic Dropdowns, and
soon to include Dynamic Dropdowns.