Update products_with_attributes_stock.php
Changed the datatype of the quantity storage variable for the updateAttribQty quantity.  The value is reaching this point as a double, but ZC does not have a double data type.  Either the value should arrive as a float (which requires some modification back before this section, or continue to use the passthru above.  Quantity has been verified to be numeric and potentially zero, so it appears that this should work fine.