Update stock_by_attr_install.php
Changed Enable Dynamic Dropdowns to default active for multiple attributes only provided the product is tracked by SBA. 
(i.e., if a product is tracked by SBA and has more than one option name associated with it then Dynamic Dropdowns are used as the default. If the product is not tracked by SBA and has more than one option name associated, then standard ZC presentation is given. If the 
product is tracked by SBA and has one option name associated with it, then the default SBA version of display is used.)