Update to support single_radioset usage
Modified the extra_cart_actions/stock_by_attributes.php file to support use of the single_radioset Dynamic Dropdown option.

One other file needs to be modified to provide the full functionality (i.e. after product is added to the cart, selection of the product from the cart should return back to the product with the selected option chosen from the list.  This requires modification of the dynamic dropdown class file supporting this functionality.